+++ packages/extension/src/worker/index.ts
@@ -26,7 +26,7 @@ emitter.on('message', async function onMessage(message: any) {
       ) as typeof import('vitest/node')
 
       const isOld = Number(vitestModule.version[0]) < 4
-      const workerName = isOld ? './workerOld.js' : './workerNew.js'
+      const workerName = isOld ? './workerLegacy.js' : './workerNew.js'
       const workerPath = pathToFileURL(join(__dirname, workerName))
       const initModule = await import(workerPath.toString())
