+++ packages/worker-legacy/src/watcher.ts
